Updated RequestAnimationFrame docs
